Draw Outline using Stencil testing

+++ lib/computational_geometry_exercise/include_dir/comp_geo_common.hpp
@@ -15,31 +15,31 @@ namespace lyh_cg {
     inline vec3 operator- (const vec3& v1, const vec3& v2);
     inline vec3 operator* (const vec3& v1, const vec3& v2);
     inline vec3 operator/ (const vec3& v1, const vec3& v2);
-    inline vec3 operator* (const vec3& v1, const double& t);
-    inline vec3 operator/ (const vec3& v1, const double& t);
-    inline vec3 operator* (const double& t, const vec3& v1);
-    inline vec3 operator/ (const double& t, const vec3& v1);
+    inline vec3 operator* (const vec3& v1, const float& t);
+    inline vec3 operator/ (const vec3& v1, const float& t);
+    inline vec3 operator* (const float& t, const vec3& v1);
+    inline vec3 operator/ (const float& t, const vec3& v1);
     class vec3 {
 
     public:
-        double e[3];
+        float e[3];
         bool isUnit;
         vec3() : e{ 0,0,0 }, isUnit(false) {};
-        vec3(double e0, double e1, double e2) : e{ e0, e1, e2 }, isUnit(false)  {};
+        vec3(float e0, float e1, float e2) : e{ e0, e1, e2 }, isUnit(false)  {};
         vec3(const vec3& v2) {
             e[0] = v2.e[0]; e[1] = v2.e[1]; e[2] = v2.e[2];
         }
-        inline double x() const { return e[0]; }
-        inline double y() const { return e[1]; }
-        inline double z() const { return e[2]; }
-        inline double r() const { return e[0]; }
-        inline double g() const { return e[1]; }
-        inline double b() const { return e[2]; }
+        inline float x() const { return e[0]; }
+        inline float y() const { return e[1]; }
+        inline float z() const { return e[2]; }
+        inline float r() const { return e[0]; }
+        inline float g() const { return e[1]; }
+        inline float b() const { return e[2]; }
         inline const vec3& operator+()const { return *this; }
         //nline  vec3& operator+(){return *this;}
         inline  vec3 operator-()const { return  vec3(-e[0], -e[1], -e[2]); }
-        inline  double operator[](int i)const { return e[i]; }
-        inline  double& operator[](int i) { return e[i]; }
+        inline  float operator[](int i)const { return e[i]; }
+        inline  float& operator[](int i) { return e[i]; }
         inline  void operator=(const vec3& v2) {
             e[0] = v2.e[0]; e[1] = v2.e[1]; e[2] = v2.e[2];
         }
@@ -56,18 +56,18 @@ namespace lyh_cg {
             *this = *this / v2; return *this;
         };
 
-        inline vec3& operator*=(const double t) {
+        inline vec3& operator*=(const float t) {
             e[0] *= t; e[1] *= t; e[2] *= t;
             return *this;
         };
-        inline vec3& operator/=(const double t) {
+        inline vec3& operator/=(const float t) {
             e[0] /= t; e[1] /= t; e[2] /= t;
             return *this;
         };
-        inline double length()const {
+        inline float length()const {
             return sqrt(e[0] * e[0] + e[1] * e[1] + e[2] * e[2]);
         }
-        inline double squared_length()const {
+        inline float squared_length()const {
             return   e[0] * e[0] + e[1] * e[1] + e[2] * e[2];
         }
         inline void make_unit_vector() {
@@ -78,7 +78,7 @@ namespace lyh_cg {
 
     };
 
-    inline double dot(const vec3& v1, const vec3& v2) {
+    inline float dot(const vec3& v1, const vec3& v2) {
         return v1.e[0] * v2.e[0] + v1.e[1] * v2.e[1] + v1.e[2] * v2.e[2];
     }
     inline vec3 cross(const vec3& a, const vec3& b) {
@@ -121,25 +121,25 @@ namespace lyh_cg {
             v1.z() / v2.z()
         );
     };
-    inline vec3 operator* (const vec3& v1, const double& t) {
+    inline vec3 operator* (const vec3& v1, const float& t) {
         return vec3(
             v1.x() * t,
             v1.y() * t,
             v1.z() * t
         );
     };
-    inline vec3 operator/ (const vec3& v1, const double& t) {
-        double f = 1.0 / t;
+    inline vec3 operator/ (const vec3& v1, const float& t) {
+        float f = 1.0 / t;
         return vec3(
             v1.x() * f,
             v1.y() * f,
             v1.z() * f
         );
     };
-    inline vec3 operator* (const double& t, const vec3& v1) {
+    inline vec3 operator* (const float& t, const vec3& v1) {
         return v1 * t;
     };
-    inline vec3 operator/ (const double& t, const vec3& v1) {
+    inline vec3 operator/ (const float& t, const vec3& v1) {
         return v1 / t;
     };
 
@@ -151,11 +151,11 @@ namespace lyh_cg {
         if ((v1 - v2).squared_length() < 1e-8)return true;
         else return false;
     }
-    bool refract(const vec3& v, const vec3& n, double n_r, vec3& refracted) {
+    bool refract(const vec3& v, const vec3& n, float n_r, vec3& refracted) {
         //default v is pointing into surface
         vec3 uv = unit_vector(v) * -1.0;
-        double dt = fmin(dot(uv, n), 1.0);
-        double D = 1.0 - n_r * n_r * (1.0 - dt * dt);
+        float dt = fmin(dot(uv, n), 1.0);
+        float D = 1.0 - n_r * n_r * (1.0 - dt * dt);
         if (D > 0.0) {
             refracted = n_r * (n * dt - uv) - n * sqrt(D);
             return true;
@@ -165,7 +165,7 @@ namespace lyh_cg {
     //alias name
     using point3 = vec3;
     using color = vec3;
-    inline double area2(const vec3& p, const vec3& q, const vec3& s) {
+    inline float area2(const vec3& p, const vec3& q, const vec3& s) {
        
         return p.x() * q.y() - p.y() * q.x()
                + q.x() * s.y() - q.y() * s.x()
